fetch/metadata/generated/appcache-manifest-sec-fetch-site.https.sub.https.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Site - redirection from HTTPS</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <body>
  <script>
  'use strict';

  /**
   * This test works through iframes instead of using the current document
   * element so that multiple subtests may be executed in parallel.
   */
  function induceRequest(path) {
    const iframe = document.createElement('iframe');
    iframe.src =
      '/fetch/metadata/resources/appcache-iframe.sub.html?manifest=' + path;
    return new Promise((resolve) => {
        addEventListener('message', function onMessage(event) {
          if (event.source !== iframe.contentWindow) {
            return;
          }
          removeEventListener('message', onMessage);
          resolve(event.data);
        });

        document.body.appendChild(iframe);
      })
      .then((message) => {
        if (message !== 'okay') {
          throw message;
        }
      })
      .then(() => iframe.remove());
  }

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return induceRequest('https://{{host}}:{{ports[https][0]}}/fetch/api/resources/redirect.py?location=http://{{host}}:{{ports[http][0]}}/fetch/api/resources/redirect.py?location=https://{{host}}:{{ports[https][0]}}' + path)
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers[''], '');
    });
  }, 'sec-fetch-site - HTTPS downgrade-upgrade');
  </script>
  </body>
</html>
